Adding support to compose 'memory'.
Now, the containers are started with a name that inform its origin (it means, the compose file that started it).
So, a new compose instance is now able to recognize such containers and load them into its context.

Instructions to load the previous started containers can be found on README file.