Add Basic Support for V2 Docker-Compose Configs
This commit adds support for V2 docker-compose configs. It does not add
any functionality included in the new style of config. Specifically, it
does not add functionality for the new Volume or Network entries. It
also doesn't handle the fact that some keys have had their positions in
the config tree moved.